+++ llvm/lib/Object/RecordStreamer.h
@@ -50,9 +50,9 @@ public:
   void EmitAssignment(MCSymbol *Symbol, const MCExpr *Value) override;
   bool EmitSymbolAttribute(MCSymbol *Symbol, MCSymbolAttr Attribute) override;
   void EmitZerofill(MCSection *Section, MCSymbol *Symbol, uint64_t Size,
-                    unsigned ByteAlignment) override;
+                    uint64_t ByteAlignment) override;
   void EmitCommonSymbol(MCSymbol *Symbol, uint64_t Size,
-                        unsigned ByteAlignment) override;
+                        uint64_t ByteAlignment) override;
   /// Record .symver aliases for later processing.
   void emitELFSymverDirective(MCSymbol *Alias,
                               const MCSymbol *Aliasee) override;
